requirements.txt
pymongo>=3.12.0
pymysql>=1.0.2
psycopg2-binary>=2.9.0
tabulate>=0.8.9
python-dateutil>=2.8.2
